+++ cluster/service_test.go
@@ -434,14 +434,14 @@ func mustNewMockTLSTransport() *mockTransport {
 }
 
 type mockDatabase struct {
-	executeFn func(er *command.ExecuteRequest) ([]*command.ExecuteResult, error)
+	executeFn func(er *command.ExecuteRequest) ([]*command.ExecuteQueryResponse, error)
 	queryFn   func(qr *command.QueryRequest) ([]*command.QueryRows, error)
 	requestFn func(rr *command.ExecuteQueryRequest) ([]*command.ExecuteQueryResponse, error)
 	backupFn  func(br *command.BackupRequest, dst io.Writer) error
 	loadFn    func(lr *command.LoadRequest) error
 }
 
-func (m *mockDatabase) Execute(er *command.ExecuteRequest) ([]*command.ExecuteResult, error) {
+func (m *mockDatabase) Execute(er *command.ExecuteRequest) ([]*command.ExecuteQueryResponse, error) {
 	return m.executeFn(er)
 }
 
@@ -471,8 +471,8 @@ func (m *mockDatabase) Load(lr *command.LoadRequest) error {
 }
 
 func mustNewMockDatabase() *mockDatabase {
-	e := func(er *command.ExecuteRequest) ([]*command.ExecuteResult, error) {
-		return []*command.ExecuteResult{}, nil
+	e := func(er *command.ExecuteRequest) ([]*command.ExecuteQueryResponse, error) {
+		return []*command.ExecuteQueryResponse{}, nil
 	}
 	q := func(er *command.QueryRequest) ([]*command.QueryRows, error) {
 		return []*command.QueryRows{}, nil