+++ http/service.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/rqlite/rqlite/v8/auth"
 	clstrPB "github.com/rqlite/rqlite/v8/cluster/proto"
-	"github.com/rqlite/rqlite/v8/command"
 	"github.com/rqlite/rqlite/v8/command/encoding"
 	"github.com/rqlite/rqlite/v8/command/proto"
+	"github.com/rqlite/rqlite/v8/command/sql"
 	"github.com/rqlite/rqlite/v8/db"
 	"github.com/rqlite/rqlite/v8/queue"
 	"github.com/rqlite/rqlite/v8/rtls"
@@ -47,7 +47,7 @@ type Database interface {
 	// to return rows. If timings is true, then timing information will
 	// be return. If tx is true, then either all queries will be executed
 	// successfully or it will as though none executed.
-	Execute(er *proto.ExecuteRequest) ([]*proto.ExecuteResult, error)
+	Execute(er *proto.ExecuteRequest) ([]*proto.ExecuteQueryResponse, error)
 
 	// Query executes a slice of queries, each of which returns rows. If
 	// timings is true, then timing information will be returned. If tx
@@ -106,7 +106,7 @@ type Cluster interface {
 	GetAddresser
 
 	// Execute performs an Execute Request on a remote node.
-	Execute(er *proto.ExecuteRequest, nodeAddr string, creds *clstrPB.Credentials, timeout time.Duration, retries int) ([]*proto.ExecuteResult, error)
+	Execute(er *proto.ExecuteRequest, nodeAddr string, creds *clstrPB.Credentials, timeout time.Duration, retries int) ([]*proto.ExecuteQueryResponse, error)
 
 	// Query performs an Query Request on a remote node.
 	Query(qr *proto.QueryRequest, nodeAddr string, creds *clstrPB.Credentials, timeout time.Duration) ([]*proto.QueryRows, error)
@@ -141,9 +141,8 @@ type StatusReporter interface {
 // DBResults stores either an Execute result, a Query result, or
 // an ExecuteQuery result.
 type DBResults struct {
-	ExecuteResult        []*proto.ExecuteResult
-	QueryRows            []*proto.QueryRows
 	ExecuteQueryResponse []*proto.ExecuteQueryResponse
+	QueryRows            []*proto.QueryRows
 
 	AssociativeJSON bool // Render in associative form
 	BlobsAsArrays   bool // Render BLOB data as byte arrays
@@ -161,9 +160,7 @@ func (d *DBResults) MarshalJSON() ([]byte, error) {
 		BlobsAsByteArrays: d.BlobsAsArrays,
 	}
 
-	if d.ExecuteResult != nil {
-		return enc.JSONMarshal(d.ExecuteResult)
-	} else if d.QueryRows != nil {
+	if d.QueryRows != nil {
 		return enc.JSONMarshal(d.QueryRows)
 	} else if d.ExecuteQueryResponse != nil {
 		return enc.JSONMarshal(d.ExecuteQueryResponse)
@@ -732,7 +729,7 @@ func (s *Service) handleLoad(w http.ResponseWriter, r *http.Request, qp QueryPar
 		queries := []string{string(b)}
 		er := executeRequestFromStrings(queries, qp.Timings(), false)
 
-		results, err := s.store.Execute(er)
+		response, err := s.store.Execute(er)
 		if err != nil {
 			if err == store.ErrNotLeader {
 				if s.DoRedirect(w, r, qp) {
@@ -741,7 +738,7 @@ func (s *Service) handleLoad(w http.ResponseWriter, r *http.Request, qp QueryPar
 			}
 			resp.Error = err.Error()
 		} else {
-			resp.Results.ExecuteResult = results
+			resp.Results.ExecuteQueryResponse = response
 		}
 		resp.end = time.Now()
 	}
@@ -1072,7 +1069,7 @@ func (s *Service) queuedExecute(w http.ResponseWriter, r *http.Request, qp Query
 			return
 		}
 	}
-	if err := command.Rewrite(stmts, !qp.NoRewriteRandom()); err != nil {
+	if err := sql.Process(stmts, !qp.NoRewriteRandom()); err != nil {
 		http.Error(w, fmt.Sprintf("SQL rewrite: %s", err.Error()), http.StatusInternalServerError)
 		return
 	}
@@ -1121,7 +1118,7 @@ func (s *Service) execute(w http.ResponseWriter, r *http.Request, qp QueryParams
 		return
 	}
 	stats.Add(numExecuteStmtsRx, int64(len(stmts)))
-	if err := command.Rewrite(stmts, !qp.NoRewriteRandom()); err != nil {
+	if err := sql.Process(stmts, !qp.NoRewriteRandom()); err != nil {
 		http.Error(w, fmt.Sprintf("SQL rewrite: %s", err.Error()), http.StatusInternalServerError)
 		return
 	}
@@ -1176,7 +1173,7 @@ func (s *Service) execute(w http.ResponseWriter, r *http.Request, qp QueryParams
 	if resultsErr != nil {
 		resp.Error = resultsErr.Error()
 	} else {
-		resp.Results.ExecuteResult = results
+		resp.Results.ExecuteQueryResponse = results
 	}
 	resp.end = time.Now()
 	s.writeResponse(w, qp, resp)
@@ -1207,7 +1204,7 @@ func (s *Service) handleQuery(w http.ResponseWriter, r *http.Request, qp QueryPa
 	// No point rewriting queries if they don't go through the Raft log, since they
 	// will never be replayed from the log anyway.
 	if qp.Level() == proto.QueryRequest_QUERY_REQUEST_LEVEL_STRONG {
-		if err := command.Rewrite(queries, qp.NoRewriteRandom()); err != nil {
+		if err := sql.Process(queries, qp.NoRewriteRandom()); err != nil {
 			http.Error(w, fmt.Sprintf("SQL rewrite: %s", err.Error()), http.StatusInternalServerError)
 			return
 		}
@@ -1300,7 +1297,7 @@ func (s *Service) handleRequest(w http.ResponseWriter, r *http.Request, qp Query
 	}
 	stats.Add(numRequestStmtsRx, int64(len(stmts)))
 
-	if err := command.Rewrite(stmts, qp.NoRewriteRandom()); err != nil {
+	if err := sql.Process(stmts, qp.NoRewriteRandom()); err != nil {
 		http.Error(w, fmt.Sprintf("SQL rewrite: %s", err.Error()), http.StatusInternalServerError)
 		return
 	}